refactor: app/index.ts を解体し bootstrap と gateway に分離
理想ディレクトリ構造への移行 Phase 5。唯一、起動経路という挙動に触るフェーズ。

これまで app/index.ts (441行) が Client の生成・login・全イベントの配線・
一部の処理本体・cron ジョブを1ファイルに抱えており、しかも「import した副作用で
login が走る」構造だったため、何がどの順で起きるのかコードから読めなかった。

- infra/discord/client.ts: Client の生成のみ(login しない)。
- bootstrap.ts: 起動シーケンスを明示。
    env 検証 -> イベント配線 -> cron 登録 -> ヘルスチェック起動 -> login
- server.ts: エントリポイント。モジュール読み込みより先に env を検証するため、
  bootstrap は動的 import する(後述)。
- gateway/: Discord イベントを features へ振り分ける層。
  旧 app/handlers/* を移設し、events.ts に client.on(...) の配線を集約。
  interaction_router / ready_handler / member_handler / reaction_handler を新設。
- jobs/stage_schedule_job.ts: 2時間毎の cron ジョブを分離。
- index.ts に埋まっていた処理本体を features へ移動:
    setEnrollmentCount   -> features/enrollment_count/
    退部ログ送信          -> features/retire_log/
    メンバー情報の同期     -> features/guild_sync/sync_member.ts
    起動時の VC 切断      -> features/voice/disconnect_from_vc.ts
    サポートスレッド作成時 -> features/support_tag/handle_thread_create.ts

これにより src/app/ が消滅し、infra/logging/send_error_logs.ts が
app/index.ts の client を参照していた最後の層違反も解消。依存方向は
gateway/registry -> features -> infra/shared/config の一方通行になった。

## 挙動が変わる点(意図的)

1. 環境変数の起動時検証を追加。DISCORD_BOT_TOKEN / DISCORD_BOT_ID / SERVER_ID /
   LOG4JS_CONFIG_PATH / VOICE_TEXT_API_KEY が欠けていれば起動前に落とす。
   VOICE_TEXT_API_KEY は読み上げ機能が import 時に assert しており、無いと
   どのみちプロセスが起動しないため必須に含めた。
   検証を import より先に行う必要があるため、server.ts では bootstrap を
   動的 import している。

2. ヘルスチェックが実態を反映するようになった。従来は常に 200 を返しており、
   Discord から切断されていても監視側は正常と判断していた。
   client.isReady() を見て 200 / 503 を返す。

## 挙動を維持した点(意図せぬ挙動だが忠実に移植)

guildMemberRemove は、退部ログチャンネルが未設定のとき early return しており、
部員数の更新まで行われていなかった。リファクタで黙って直すと差分が紛れるため、
sendRetireLog の戻り値で同じ挙動を再現し、コメントを残した。別途修正を検討する。

## 検証

- ダミー環境変数での起動スモークテスト:
  必須変数が欠けている場合は起動前に FATAL で停止。
  全変数が揃っている場合はイベント配線・cron 登録・ヘルスチェック起動まで進み、
  不正トークンで login に失敗して停止することを確認。
- エントリポイントは build/server.js のまま維持(systemd の起動設定に追随不要)。
- lint・型チェック・全80テスト・ビルド green。
