+++ src/gateway/button_handler.ts
