+++ src/gateway/command_handler.ts
