+++ src/gateway/context_handler.ts
