+++ src/gateway/message_handler.ts
