+++ src/gateway/modal_handler.ts
