+++ src/gateway/voice_state_handler.ts
